Initial rewrite for version 2
The commit marks to start of version 2.0.0 of the SQL LocalDB API wrapper.

It is effectively a rewrite, and is a breaking change to upgrade from 1.x.x.

The main changes involve support for .NET Standard 2.0, as well as redesign to
better support principles such as dependency injection and extensibility.

  * All user-facing static types (except for extensions) have been removed.
  * The NuGet package and the namespaces have been renamed to not be in System.
  * Configuration is now via classes, not System.Configuration support.
  * The solution is now compiled using the .NET Core SDK.
  * Tests are now written in xunit instead of MSTest.
  * All extensions to parse connection strings for EntityFramework are removed.
  * Updates to source code to use C# 7.2 features where appropriate.
  * Various method overloads of old types are now extensions for new types.